requirements.txt
Flask
numpy
faiss-cpu
PyMuPDF
scikit-learn
pdfplumber